Project Overview
In this we build the full process of building a machine learning project. This includes creating a hypothesis, setting up the model, and measuring error  using Python and Jupyter.

To make this interesting, we'll use a fun dataset. We'll use data from historical Olympic games. We'll try to predict how many medals a country will win based on historical and current data.

Project Steps

Form a hypothesis.
Find and explore the data.
(If necessary) Reshape the data to predict your target.
Clean the data for ML.
Pick an error metric.
Split your data.
Train a model.

To follow this project, please install the following locally:

Python 3.8+
Python packages
pandas
numpy
scikit-learn
seaborn